[wip] detect if nodes are in the same region; refoactor code
